refactor(search): unify /all + /discover on inline suggestions, drop search-box-ux
Extract the discover-page two-mode dropdown into a shared
InlineSearchSuggestions component and mount it on /all's HomeHeading
so both pages share the same search UX (keyword origin/resource
suggestions pre-space, semantic LLM results post-space).

On /all there's no body-level results view, so submitting from there
hands off to the discover page via /?q=… which renders the full
search results table.

Delete the entire search-box-ux feature now that it has no consumers:
- apps/scan/src/features/search-box/* (~2,000 lines)
- apps/scan/src/app/api/search/* proxy routes
- apps/scan/src/app/(app)/_components/search/x402-linked-search-box.tsx
- AGENTCASH_SEARCH_API_KEY + AGENTCASH_SEARCH_API_URL env vars

The discover semantic search uses a different upstream
(AGENTCASH_URL + AGENTCASH_INTERNAL_API_KEY → /api/internal/search),
which is unaffected.

Inline rows use <Link> instead of <button> so the inner Tooltip
trigger button inside Origin/Addresses doesn't violate
button-in-button hydration.
