+++ apps/scan/src/app/(app)/(home)/(discover)/_components/discover-search.tsx
@@ -1,7 +1,6 @@
 'use client';
 
-import { CornerDownLeft, Loader2, Plus, Search, X } from 'lucide-react';
-import { useRouter } from 'next/navigation';
+import { CornerDownLeft, Plus, Search, X } from 'lucide-react';
 import { useState } from 'react';
 
 import { Button } from '@/components/ui/button';
@@ -9,10 +8,7 @@ import { Input } from '@/components/ui/input';
 import { useChain } from '@/app/(app)/_contexts/chain/hook';
 import { useSellersSorting } from '@/app/(app)/_contexts/sorting/sellers/hook';
 import { api } from '@/trpc/client';
-import { Origin } from '@/app/(app)/_components/origins';
-import { Resource } from '@/app/(app)/_components/resource';
-import { Favicon } from '@/app/(app)/_components/favicon';
-import { useDebounce } from '@/hooks/use-debounce';
+import { InlineSearchSuggestions } from '@/app/(app)/_components/search/inline-search-suggestions';
 
 import { DataTable } from '@/components/ui/data-table';
 import { discoverColumns } from './columns';
@@ -72,159 +68,17 @@ export const DiscoverSearchInput = () => {
   );
 };
 
-/**
- * Live dropdown shown beneath the discover search input. Pre-space we hit the
- * same origin/resource search the navbar uses; once the user types a space
- * (signaling a phrase) we swap to the semantic search endpoint.
- */
 const DiscoverInlineSuggestions: React.FC<{ enabled: boolean }> = ({
   enabled,
 }) => {
-  const router = useRouter();
   const { input, query, submit } = useDiscoverSearch();
-  const trimmed = input.trim();
-  // Only show while the user is editing — once the query is committed and the
-  // body view takes over, hide the dropdown to avoid double-rendering matches.
-  const isDirty = trimmed !== query;
-  const isVisible = enabled && trimmed.length > 0 && isDirty;
-  // A space anywhere in the live input promotes from keyword → semantic mode.
-  const isSemanticMode = input.includes(' ');
-
-  // Real debounce on the LLM-backed semantic search — useDeferredValue only
-  // batches React state, it doesn't suppress requests, so rapid typing in
-  // semantic mode would still hammer the upstream LLM. Cheap keyword queries
-  // (origins/resources) stay live since they're DB-fast.
-  const debouncedSemanticInput = useDebounce(
-    isSemanticMode ? trimmed : '',
-    250
-  );
-
-  const originSearch = api.public.origins.search.useQuery(
-    { search: trimmed, limit: 5 },
-    { enabled: isVisible && !isSemanticMode }
-  );
-  const resourceSearch = api.public.resources.search.useQuery(
-    { search: trimmed, limit: 5 },
-    { enabled: isVisible && !isSemanticMode }
-  );
-  const semanticSearch = api.public.discover.search.useQuery(
-    { query: debouncedSemanticInput },
-    {
-      enabled: isVisible && isSemanticMode && debouncedSemanticInput.length > 0,
-    }
-  );
-
-  if (!isVisible) return null;
-
-  const origins = originSearch.data ?? [];
-  const resources = resourceSearch.data ?? [];
-  const semanticResults = semanticSearch.data ?? [];
-  const keywordLoading = originSearch.isLoading || resourceSearch.isLoading;
-  const semanticLoading =
-    semanticSearch.isLoading || trimmed !== debouncedSemanticInput;
-
   return (
-    <div className="absolute top-full left-0 right-0 mt-1 z-50 rounded-md border bg-popover shadow-lg overflow-hidden">
-      {isSemanticMode ? (
-        semanticLoading ? (
-          <div className="flex items-center gap-2 px-3 py-3 text-sm text-muted-foreground">
-            <Loader2 className="size-4 animate-spin" />
-            Searching x402 services for &quot;{trimmed}&quot;…
-          </div>
-        ) : semanticResults.length > 0 ? (
-          <ul className="max-h-80 overflow-y-auto">
-            <li className="px-3 pt-2 pb-1 text-[11px] font-medium uppercase tracking-[0.18em] text-muted-foreground">
-              Best matches
-            </li>
-            {semanticResults.map(result => (
-              <li key={result.origin}>
-                <button
-                  type="button"
-                  onMouseDown={e => e.preventDefault()}
-                  onClick={() => submit()}
-                  className="w-full px-3 py-2 flex items-center gap-3 hover:bg-accent text-left"
-                >
-                  <Favicon url={result.favicon} className="size-6 shrink-0" />
-                  <div className="min-w-0 flex-1">
-                    <div className="truncate text-sm font-medium">
-                      {result.title || new URL(result.origin).hostname}
-                    </div>
-                    {result.description ? (
-                      <div className="truncate text-xs text-muted-foreground">
-                        {result.description}
-                      </div>
-                    ) : null}
-                  </div>
-                </button>
-              </li>
-            ))}
-          </ul>
-        ) : (
-          <div className="px-3 py-3 text-sm text-muted-foreground">
-            No semantic matches. Press Enter to refine.
-          </div>
-        )
-      ) : keywordLoading ? (
-        <div className="flex items-center gap-2 px-3 py-3 text-sm text-muted-foreground">
-          <Loader2 className="size-4 animate-spin" />
-          Searching origins…
-        </div>
-      ) : origins.length === 0 && resources.length === 0 ? (
-        <div className="px-3 py-3 text-sm text-muted-foreground">
-          No matching origins or resources. Add a space to search by phrase.
-        </div>
-      ) : (
-        <ul className="max-h-80 overflow-y-auto">
-          {origins.length > 0 ? (
-            <>
-              <li className="px-3 pt-2 pb-1 text-[11px] font-medium uppercase tracking-[0.18em] text-muted-foreground">
-                Origins
-              </li>
-              {origins.map(origin => (
-                <li key={`origin-${origin.id}`}>
-                  <button
-                    type="button"
-                    onMouseDown={e => e.preventDefault()}
-                    onClick={() => router.push(`/server/${origin.id}`)}
-                    className="w-full px-3 py-2 hover:bg-accent text-left"
-                  >
-                    <Origin
-                      origin={origin}
-                      addresses={Array.from(
-                        new Set(
-                          origin.resources.flatMap(resource =>
-                            resource.accepts.map(accept => accept.payTo)
-                          )
-                        )
-                      )}
-                    />
-                  </button>
-                </li>
-              ))}
-            </>
-          ) : null}
-          {resources.length > 0 ? (
-            <>
-              <li className="px-3 pt-2 pb-1 text-[11px] font-medium uppercase tracking-[0.18em] text-muted-foreground">
-                Resources
-              </li>
-              {resources.map(resource => (
-                <li key={`resource-${resource.id}`}>
-                  <button
-                    type="button"
-                    onMouseDown={e => e.preventDefault()}
-                    onClick={() => router.push(`/server/${resource.origin.id}`)}
-                    className="w-full px-3 py-2 hover:bg-accent text-left"
-                  >
-                    <Resource resource={resource} />
-                  </button>
-                </li>
-              ))}
-            </>
-          ) : null}
-        </ul>
-      )}
-    </div>
+    <InlineSearchSuggestions
+      input={input}
+      enabled={enabled}
+      committedQuery={query}
+      onSubmit={submit}
+    />
   );
 };
