--- apps/scan/src/app/api/search/autocomplete/route.ts
@@ -1,11 +1,0 @@
-import { proxyAgentCashSearchGet } from '../_lib/agentcash-search-proxy';
-
-const autocompleteParams = ['q', 'suggestions', 'results'] as const;
-
-export function GET(request: Request) {
-  return proxyAgentCashSearchGet({
-    allowedParams: autocompleteParams,
-    path: '/api/external/search/autocomplete',
-    request,
-  });
-}
