--- apps/scan/src/app/api/search/events/feedback/route.ts
@@ -1,8 +1,0 @@
-import { proxyAgentCashSearchPost } from '../../_lib/agentcash-search-proxy';
-
-export function POST(request: Request) {
-  return proxyAgentCashSearchPost({
-    path: '/api/external/search/events/feedback',
-    request,
-  });
-}
