--- apps/scan/src/app/api/search/events/selection/route.ts
@@ -1,8 +1,0 @@
-import { proxyAgentCashSearchPost } from '../../_lib/agentcash-search-proxy';
-
-export function POST(request: Request) {
-  return proxyAgentCashSearchPost({
-    path: '/api/external/search/events/selection',
-    request,
-  });
-}
