--- apps/scan/src/app/api/search/results/route.ts
@@ -1,11 +1,0 @@
-import { proxyAgentCashSearchGet } from '../_lib/agentcash-search-proxy';
-
-const searchResultParams = ['q', 'broad', 'limit'] as const;
-
-export function GET(request: Request) {
-  return proxyAgentCashSearchGet({
-    allowedParams: searchResultParams,
-    path: '/api/external/search/results',
-    request,
-  });
-}
