--- apps/scan/src/features/search-box/index.ts
@@ -1,8 +1,0 @@
-export { SearchBox } from './search-box';
-export type {
-  AutocompleteSuggestion,
-  SearchBoxFeedbackContext,
-  SearchBoxSelectionContext,
-  SearchPreviewResult,
-} from './types';
-export type { SearchBoxProps } from './search-box';
